Klasa tekstury automatycznie włącza odpowiedni caps OpenGL. Dodanie komponentów jednostek: Movable, StaticImage. Rozszerzenie klasy SampleWorkerDescription o dodane komponenty. Klasa stanu gry testowo dodaje nowo utworzoną jednostkę. Kilka mniejszych zmian.

+++ Src/Kingdoms Clash.NET/Interfaces/Units/Components/IMovable.cs
@@ -7,8 +7,8 @@
 		: IUnitComponent
 	{
 		/// <summary>
-		/// Szybkość wyrażona w jednostce j/sec.
+		/// Szybkość poruszania się jednostki.
 		/// </summary>
-		int Speed { get; }
+		float Speed { get; }
 	}
 }
